mqtt

+++ Core/Src/Peripheral/ten.h
@@ -13,6 +13,6 @@
 uint8_t ten_power;
 
 void ten_handler();
-void set_ten_power(uint8_t value);
+void set_ten_power(uint16_t value);
 
 #endif // __TEN_H__
